// web/js/merchant/views/Settings/Configuration/CheckoutEditor/context/types.d.ts
export type CheckoutEditorContext<Values> = {
  values: Values;
  isValueModified: boolean;
  isSaving: boolean;
  isLoading: boolean;
  handleSave: () => void;
  handleLocaleChange: (value: string) => void;
  handleDiscardAllChanges: () => void;
  handleCustomMessageToggle: (isEnabled: boolean) => void;
  handleCustomMessageTextChange: (index: number, value: string) => void;
  handleCustomMessageTextColorChange: (index: number, value: string) => void;
  handleCustomMessageBackgroundColorChange: (index: number, value: string) => void;
  handlePreviewChange: (value: boolean) => void;
  handleFlashCheckoutToggle: (isEnabled: boolean) => void;
  handleEmailValueChange: (value: string) => void;
  handleEmailToggle: (isEnabled: boolean) => void;
  handleMandatorySummaryPageToggle: (isEnabled: boolean) => void;
  handleShowFinalPriceToggle: (isEnabled: boolean) => void;
  handleLogoChange: (file: File | null) => void;
  handleRectLogoChange: (file: File | null) => void;
  handleBrandColorChange: (evt?: React.ChangeEvent) => void;
  handleButtonStyleChange: (value: string) => void;
  handleFontStyleChange: (value: string) => void;
  handleSidebarGraphicToggle: (value: boolean) => void;
  handleSidebarGraphicValueChange: (value: string) => void;
};